Create build and basic test suite with GH actions
Adding a test suite that runs when the listed branches are pushed to,
or when a pull-request is opened against main.
The list of specific branches for either of these tests can be
replaced with "**" to instead run on all branches.
A selection of tests have been chosen to act as a blueprint for testing; the remainder of tests in MesmerQA can be added using the
current tests as a template.
Only "Acetyl O2 association" and "Unit tests" are repeated on the
parallel build; the other tests can be copied across too if wished.
Note that tests which require different baselines for newer versions
of Linux have had "NEW" added to the end of the test name. These are
compared against new baselines generated on Ubuntu 20.04.
The tests will fail on initial run because new baselines have not
yet been added to the repository.

Co-authored-by: John Hodrien <johnhodrien@gmail.com>